🚚 Rename: 폴더명 오타 수정
