+++ src/components/workdetail/rightcomponent/RightContent.jsx
